+++ src/app/[locale]/(logged-out)/about/cohort-discovery-request/page.tsx
@@ -4,7 +4,7 @@ import ProtectedAccountRoute from "@/components/ProtectedAccountRoute";
 import { getUserFromCookie } from "@/utils/api";
 import { getCohortTermsAndConditions } from "@/utils/cms";
 import metaData from "@/utils/metadata";
-import CohortDisoveryRequestForm from "./components/CohortDisoveryRequestForm";
+import CohortDiscoveryRequestForm from "./components/CohortDiscoveryRequestForm";
 
 export const metadata = metaData({
     title: "Cohort Discovery Request - About",
@@ -23,7 +23,7 @@ export default async function CohortDiscoryRequestPage() {
     return (
         <ProtectedAccountRoute loggedInOnly={!!user?.id}>
             <Container>
-                <CohortDisoveryRequestForm
+                <CohortDiscoveryRequestForm
                     userId={user?.id}
                     cmsContent={repeatfields}
                 />